requirements.txt
openai
python-dotenv
pywebview
